Remove unused classes
