+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/system/SystemController.java
@@ -18,7 +18,12 @@
 
 package com.radixdlt.api.system;
 
+import com.radixdlt.DefaultSerialization;
+import com.radixdlt.client.ValidatorAddress;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
+import com.radixdlt.network.addressbook.AddressBook;
+import com.radixdlt.network.addressbook.PeerWithSystem;
+import com.radixdlt.serialization.DsonOutput;
 import com.radixdlt.statecomputer.checkpoint.Genesis;
 import com.radixdlt.systeminfo.InMemorySystemInfo;
 import com.radixdlt.utils.Bytes;
@@ -31,23 +36,32 @@ import com.google.inject.Inject;
 
 import io.undertow.server.HttpServerExchange;
 import io.undertow.server.RoutingHandler;
+import org.radix.universe.system.LocalSystem;
 
+import java.util.stream.Stream;
+
+import static com.radixdlt.api.JsonRpcUtil.jsonObject;
 import static com.radixdlt.api.RestUtils.respond;
 
 public final class SystemController implements Controller {
-	private final SystemService systemService;
+	private final LocalSystem localSystem;
 	private final VerifiedTxnsAndProof genesis;
 	private final InMemorySystemInfo inMemorySystemInfo;
+	private final AddressBook addressBook;
+	private final PeerWithSystem localPeer;
 
 	@Inject
 	public SystemController(
 		InMemorySystemInfo inMemorySystemInfo,
-		SystemService systemService,
+		LocalSystem localSystem,
+		AddressBook addressBook,
 		@Genesis VerifiedTxnsAndProof genesis
 	) {
 		this.inMemorySystemInfo = inMemorySystemInfo;
-		this.systemService = systemService;
+		this.addressBook = addressBook;
 		this.genesis = genesis;
+		this.localSystem = localSystem;
+		this.localPeer = new PeerWithSystem(this.localSystem);
 	}
 
 	@Override
@@ -56,6 +70,7 @@ public final class SystemController implements Controller {
 		handler.get("/system/checkpoints", this::respondWithGenesis);
 		handler.get("/system/proof", this::respondWithCurrentProof);
 		handler.get("/system/epochproof", this::respondWithEpochProof);
+		handler.get("/system/peers", this::respondWithLivePeers);
 	}
 
 	void respondWithCurrentProof(final HttpServerExchange exchange) {
@@ -70,7 +85,8 @@ public final class SystemController implements Controller {
 
 	@VisibleForTesting
 	void respondWithLocalSystem(final HttpServerExchange exchange) {
-		respond(exchange, systemService.getLocalSystem());
+		var json = DefaultSerialization.getInstance().toJsonObject(localSystem, DsonOutput.Output.API);
+		respond(exchange, json);
 	}
 
 	@VisibleForTesting
@@ -82,5 +98,28 @@ public final class SystemController implements Controller {
 		jsonObject.put("proof", genesis.getProof().asJSON());
 
 		respond(exchange, jsonObject);
+	}
+
+
+	private Stream<PeerWithSystem> selfAndOthers(Stream<PeerWithSystem> others) {
+		return Stream.concat(Stream.of(this.localPeer), others).distinct();
+	}
+
+	private void respondWithLivePeers(final HttpServerExchange exchange) {
+		var peerArray = new JSONArray();
+		selfAndOthers(this.addressBook.recentPeers())
+			.map(peer -> {
+				var json = jsonObject().put("address", ValidatorAddress.of(peer.getSystem().getKey()));
+				peer.getSystem().supportedTransports().filter(t -> t.name().equals("TCP")).forEach(t -> {
+					String port = t.metadata().get("port");
+					String host = t.metadata().get("host");
+					json.put("endpoint", host + ":" + port);
+				});
+
+				return json;
+			})
+			.forEach(peerArray::put);
+
+		respond(exchange, peerArray);
 	}
 }
