--- radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/system/SystemService.java
@@ -1,44 +1,0 @@
-/*
- * (C) Copyright 2021 Radix DLT Ltd
- *
- * Radix DLT Ltd licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except in
- * compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied.  See the License for the specific
- * language governing permissions and limitations under the License.
- *
- */
-
-package com.radixdlt.api.system;
-
-import org.json.JSONObject;
-import org.radix.universe.system.LocalSystem;
-
-import com.google.inject.Inject;
-import com.radixdlt.serialization.DsonOutput;
-import com.radixdlt.serialization.Serialization;
-
-public class SystemService {
-	private final Serialization serialization;
-	private final LocalSystem localSystem;
-
-	@Inject
-	public SystemService(
-		Serialization serialization,
-		LocalSystem localSystem
-	) {
-		this.serialization = serialization;
-		this.localSystem = localSystem;
-	}
-
-	public JSONObject getLocalSystem() {
-		return serialization.toJsonObject(localSystem, DsonOutput.Output.API);
-	}
-}
