--- radixdlt-core/radixdlt/src/test/java/org/radix/api/services/SystemServiceTest.java
@@ -1,95 +1,0 @@
-/*
- * (C) Copyright 2021 Radix DLT Ltd
- *
- * Radix DLT Ltd licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except in
- * compliance with the License.  You may obtain a copy of the
- * License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied.  See the License for the specific
- * language governing permissions and limitations under the License.
- */
-
-package org.radix.api.services;
-
-import com.radixdlt.api.system.SystemService;
-import org.junit.Test;
-import org.radix.universe.system.LocalSystem;
-
-import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableList;
-import com.radixdlt.DefaultSerialization;
-import com.radixdlt.crypto.ECKeyPair;
-import com.radixdlt.network.transport.StaticTransportMetadata;
-import com.radixdlt.network.transport.TransportInfo;
-import com.radixdlt.network.transport.udp.UDPConstants;
-import com.radixdlt.serialization.Serialization;
-import com.radixdlt.utils.Bytes;
-
-import java.io.IOException;
-import java.util.Map;
-import java.util.Properties;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.radix.Radix.AGENT;
-import static org.radix.Radix.AGENT_VERSION;
-import static org.radix.Radix.PROTOCOL_VERSION;
-
-public class SystemServiceTest {
-	private static final String PROPERTIES_FILE = "/default.config";
-
-	private final Serialization serialization = DefaultSerialization.getInstance();
-	private final LocalSystem localSystem = assembleLocalSystem();
-	private final SystemService systemService = new SystemService(serialization, localSystem);
-
-	@Test
-	public void localSystemCanBeObtained() {
-		var result = systemService.getLocalSystem();
-
-		assertNotNull(result);
-
-		var agent = result.getJSONObject("agent");
-
-		assertEquals(PROTOCOL_VERSION, agent.get("protocol"));
-		assertEquals(":str:" + AGENT, agent.getString("name"));
-		assertEquals(AGENT_VERSION, agent.get("version"));
-	}
-
-	private static String universeBase64() throws IOException {
-		final var properties = new Properties();
-
-		try (final var input = SystemService.class.getResourceAsStream(PROPERTIES_FILE)) {
-			if (input == null) {
-				throw new IOException("Resource not found: " + PROPERTIES_FILE);
-			}
-			properties.load(input);
-		}
-		return (String) properties.get("universe");
-	}
-
-	private static LocalSystem assembleLocalSystem() {
-		try {
-			var keyPair = ECKeyPair.fromPrivateKey(Bytes.fromHexString(Strings.repeat("deadbeef", 8)));
-
-			var transports = ImmutableList.of(
-				TransportInfo.of(
-					UDPConstants.NAME,
-					StaticTransportMetadata.of(
-						UDPConstants.METADATA_HOST, "127.0.0.1",
-						UDPConstants.METADATA_PORT, "30000"
-					)
-				)
-			);
-
-			return new LocalSystem(Map::of, keyPair.getPublicKey(), AGENT, AGENT_VERSION, PROTOCOL_VERSION, transports);
-		} catch (Exception e) {
-			throw new RuntimeException("Unable to create local system", e);
-		}
-	}
-}
